### 2021-02-17 #3043 @whedon
[PRE REVIEW]: SkyPy: A package for modelling the Universe
**Submitting author:** @rrjbca (<a href="http://orcid.org/0000-0003-1291-1023">Richard Rollins</a>)
**Repository:** <a href="https://github.com/skypyproject/skypy" target ="_blank">https://github.com/skypyproject/skypy</a>
**Version:** v0.4
