### 2021-02-17 #3043 @arfon
> @openjournals/dev can you help trouble shoot this error ☝️

I think the `@` in the emails in the authors need escaping, e.g.:

```
authors:
  - name: Adam Amara^[adam.amara\@port.ac.uk]
    orcid: 0000-0003-3481-3491
    affiliation: 1
```
