### 2021-02-17 #3043 @Kevin-Mattheus-Moerman
@rrjbca can you implement these changes and call `@whedon generate pdf from branch joss-paper` here to update the paper?
