### 2021-02-17 21a7285
Fix flake8 code style errors in power_spectrum module (#430)
