+++ skypy/power_spectrum/tests/test_import.py
@@ -1,2 +1,2 @@
 def test_import():
-    import skypy.power_spectrum
+    import skypy.power_spectrum  # noqa: F401
