### 2021-02-17 862af0b
Fix flake8 code style errors in halos module (#431)
