+++ skypy/halos/tests/test_import.py
@@ -1,5 +1,5 @@
 def test_import():
-    import skypy.halos
-    import skypy.halos.abundance_matching
-    import skypy.halos.mass
-    import skypy.halos.quenching
+    import skypy.halos  # noqa: F401
+    import skypy.halos.abundance_matching  # noqa: F401
+    import skypy.halos.mass  # noqa: F401
+    import skypy.halos.quenching  # noqa: F401
