### 2021-02-18 dd4f78c
Update skypy/clusters/sampler.py
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
