### 2021-02-18 2f06a8e
move sampler.py to halos folder
