### 2021-02-18 5da23e7
remove selection function, add resolution argument
