### 2021-02-19 8376d6f
update docs/halos/index.rst
