+++ docs/halos/index.rst
@@ -109,7 +109,7 @@ Quenching (`skypy.halos.quenching`)
 .. automodule:: skypy.halos.quenching
 
 
-Quenching (`skypy.halos.sampler`)
+Sampler (`skypy.halos.sampler`)
 ===================================
 
 .. automodule:: skypy.halos.sampler
