+++ .github/workflows/compatibility.yaml
@@ -11,13 +11,13 @@ jobs:
         os: [ubuntu-latest]
         python: [3.9]
         toxenv: [py39-test, py39-test-dev]
-        release: [master, latest]
+        release: [main, latest]
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v2
       with:
         fetch-depth: 0
-    - if: matrix.release != 'master'
+    - if: matrix.release != 'main'
       name: Checkout Release
       run: |
         git checkout tags/$(curl -s https://api.github.com/repos/skypyproject/skypy/releases/${{ matrix.release }} | python -c "import sys, json; print(json.load(sys.stdin)['tag_name'])")
