### 2021-02-21 #3043 @arfon
@rrjbca - thanks for your submission to JOSS? Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission? (Please avoid @mentioning them here).

Also, if there are other people in the community who you think might be able to give this package a good review, please let me know.
