+++ CONTRIBUTING.rst
@@ -1,11 +1,58 @@
 Contributor Guidelines
 ======================
 
-GitHub Workflow
----------------
+Developing new features
+-----------------------
 
-Fork and Clone the SkyPy Repository
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+To develop new features for SkyPy, you will normally want to create a local copy of the SkyPy repository, instead of installing e.g. a pip or conda package.
+
+
+Clone the SkyPy Repository
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Cloning creates a local copy of the repository on your computer. To clone the SkyPy repository, navigate to where you keep your code files in a terminal, and run `git clone`::
+
+  # clone the repository ...
+  git clone https://github.com/skypyproject/skypy.git
+  # ... which will create the `skypy` folder for you
+  cd skypy
+
+
+Create a branch for your new feature
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Create a *branch* off the ``skypyproject`` main branch. Working on unique branches for each new feature simplifies the development, review, and merge processes by maintining logical separation. To create a feature branch::
+
+  # first, make sure we are on the main branch
+  git checkout main
+  # pull in potential changes from the repository
+  git pull
+  # now create your new branch with name "BRANCH_NAME"
+  git checkout -b BRANCH_NAME
+
+
+Hack away!
+^^^^^^^^^^
+
+Write the new code you would like to contribute and *commit* it to the feature branch on your local repository. Ideally commit small units of work often, with clear and descriptive commit messages describing the changes you made. To commit changes to a file::
+
+  # see the changes in the working directory
+  git status
+  # add a file called "FILE_WITH_CHANGES" to the index to be committed
+  git add FILE_WITH_CHANGES
+  # commit all staged changes from the index
+  git commit
+
+
+Contributing
+------------
+
+Once you have developed your new SkyPy feature, you may want to contribute it to the main repository. To do so, you have to create your own copy of the repository (called a "fork") once, push your changes there, and then propose them for merging into the main repository. Please keep the coding guidelines below in mind.
+
+
+Fork the SkyPy Repository
+^^^^^^^^^^^^^^^^^^^^^^^^^
+
 **You should only need to do this step once**
 
 First *fork* the SkyPy repository. A fork is your own remote copy of the repository on GitHub. To create a fork:
@@ -14,69 +61,40 @@ First *fork* the SkyPy repository. A fork is your own remote copy of the reposit
   2. Click the **Fork** button (in the top-right-hand corner)
   3. Choose where to create the fork, typically your personal GitHub account
 
-Next *clone* your fork. Cloning creates a local copy of the repository on your computer to work with. To clone your fork:
 
-::
-
-   git clone https://github.com/<your-account>/skypy.git
-
-
-Finally add the ``skypyproject`` repository as a *remote*. This will allow you to fetch changes made to the codebase. To add the ``skypyproject`` remote:
-
-::
-
-  cd skypy
-  git remote add skypyproject https://github.com/skypyproject/skypy.git
-
-
-Create a branch for your new feature
+Add your fork as a remote repository
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Create a *branch* off the ``skypyproject`` master branch. Working on unique branches for each new feature simplifies the development, review and merge processes by maintining logical separation. To create a feature branch:
+Next, you need to make your fork known to your local SkyPy repository. To add a *remote* repository::
 
-::
+  git remote add GITHUB_USER https://github.com/GITHUB_USER/skypy.git
 
-  git fetch skypyproject
-  git checkout -b <your-branch-name> skypyproject/master
-
-
-Hack away!
-^^^^^^^^^^
-
[…]
