### 2021-02-22 #3043 @arfon
:wave: @cescalara @rmorgan10 @adrn - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is [_SkyPy: A package for modelling the Universe_](https://github.com/skypyproject/skypy)
