### 2021-02-22 #3043 @arfon
@rrjbca - does this submission accompany a paper in a AAS journal?
