### 2021-02-23 #3043 @rrjbca
> @rrjbca - does this submission accompany a paper in a AAS journal?

@arfon No this is a stand-alone submission
