### 2021-02-23 5ae72fc
Update docs/arch/adr-03.md
