### 2021-02-24 3a5987c
Update skypy/halos/sampler.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
