### 2021-02-24 46fede8
Update skypy/halos/sampler.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
