+++ skypy/power_spectrum/_camb.py
@@ -85,11 +85,16 @@ def camb(wavenumber, redshift, cosmology, A_s, n_s):
 
     pars.NonLinear = model.NonLinear_none
 
+    if len(redshift) > 1:
+        redshift_kwargs = {'zmin': np.min(redshift), 'zmax': np.max(redshift)}
+    else:
+        redshift_kwargs = {'zs': redshift}
+
     pk_interp = get_matter_power_interpolator(pars,
                                               nonlinear=False,
                                               hubble_units=False, k_hunit=False,
                                               kmax=np.max(wavenumber),
-                                              zmax=np.max(redshift))
+                                              **redshift_kwargs)
 
     pzk = pk_interp.P(redshift[redshift_order], wavenumber[wavenumber_order])
 
