### 2021-02-26 4db14a8
Revert "remove module dependencies (#392)"
This reverts commit 225de6fd63142ef6d6d3e1be3b0cb81af6fea073.
