+++ docs/conf.py
@@ -251,6 +251,13 @@ plot_rcparams = {
     'text.usetex': False,
 }
 
+# Mock camb if it is not available e.g. on readthedocs
+autodoc_mock_imports = []
+try:
+    import_module('camb')
+except ImportError:
+    autodoc_mock_imports.append('camb')
+
 # Sphinx Gallery
 extensions += ['sphinx_gallery.gen_gallery', ]
 sphinx_gallery_conf = {
