+++ setup.cfg
@@ -33,6 +33,8 @@ test =
     pytest-rerunfailures
     speclite>=0.11
 all =
+    camb
+    classy==2.8.*
     speclite>=0.11
 docs =
     sphinx-astropy
