### 2021-03-01 d43ec37
Optimise for redshift arrays smaller than nz_step
