+++ skypy/power_spectrum/_camb.py
@@ -1,5 +1,6 @@
 import numpy as np
 from astropy import units
+from inspect import signature
 
 
 __all__ = [
@@ -85,11 +86,16 @@ def camb(wavenumber, redshift, cosmology, A_s, n_s):
 
     pars.NonLinear = model.NonLinear_none
 
+    if len(redshift) > signature(get_matter_power_interpolator).parameters['nz_step'].default:
+        redshift_kwargs = {'zmin': np.min(redshift), 'zmax': np.max(redshift)}
+    else:
+        redshift_kwargs = {'zs': redshift}
+
     pk_interp = get_matter_power_interpolator(pars,
                                               nonlinear=False,
                                               hubble_units=False, k_hunit=False,
                                               kmax=np.max(wavenumber),
-                                              zmax=np.max(redshift))
+                                              **redshift_kwargs)
 
     pzk = pk_interp.P(redshift[redshift_order], wavenumber[wavenumber_order])
 
