### 2021-03-02 skypyproject/skypy#443 @rmorgan10
Comprehensive List of Features - JOSS Review
After some time digging through the code and documentation, I was able to piece together that SkyPy is useful for quickly obtaining realistic distributions of galaxy shapes, redshifts, luminosities, photometry, and masses (perhaps there are more physical quantities that I missed). If that list of features / use cases were clearly presented, it would eliminate the need for interested astronomers to dig through all the code / examples / documentation to determine whether SkyPy will be useful in their analysis.

I recommend 2 ways of conveying all of SkyPy's uses to potential users:

1. In the software paper, include the physical quantities related to SkyPy in the high-level description of the package's functionality

and

2. In the documentation, create a new page titled "Main Functionalities" or something to that extent. In your main README, link to this specific page in the sentence "The full list of features ...". In the new page, link to the location in the documentation for each feature you list.

Examples:
[Galaxy Morphology Distributions](https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.morphology)
[Galaxy Redshift Distributions](https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.redshift)

The page https://skypy.readthedocs.io/en/latest/galaxies.html _almost_ accomplishes this goal for the skypy.galaxies package, but the only descriptions included are "The following modules are found in the skypy.galaxies.* package." I think a potential user may want some expanded descriptions of how a particular module will be relevant to their analysis.

The page I think would be useful would contain all packages in SkyPy (at present galaxies, utils, and pipeline) and use "astronomer-friendly" language (not function names) to describe the feature, and then link to the page in the documentation. All the features are documented thoroughly, but it would be nice to have some sort of reference page for all the things SkyPy can do.

After seeing some of the other issues in this repo, it's clear that the main functionalities of SkyPy will be expanding, so that is even stronger motivation to have a page like this.
