### 2021-03-02 skypyproject/skypy#444 @rmorgan10
Guide to Constructing Configuration Files - JOSS Review
To make getting up and running as painless as possible for new users, I recommend a guide on how to go about writing a general configuration file. Clearly (based on the examples) there is a natural flow to the configuration file. It starts at the top with definitions of individual variables, then goes into the tables to produce, and within the table entries the characteristics of the objects to simulate are included. A "User Guide" page in the documentation could certainly provide a walkthrough of this flow and how to think through the process of creating a general configuration file.

Furthermore, a specific request I have for this sort of guide is to list all the filters that can be used. I was unable to find what the other options instead of "sdss2010-*" were in the galaxy photometry example. If you could point out where this documentation is (if I just missed it) that would also work for me.
