### 2021-03-02 skypyproject/skypy#443 @rmorgan10
Linking https://github.com/openjournals/joss-reviews/issues/3056
