### 2021-03-03 35524f2
Fixed wavenumber definition in E&Hu model
