### 2021-03-04 c2b938c
add unit tests and documentation
