+++ skypy/halos/sampler.py
@@ -1,17 +1,33 @@
-'''Colossus halo mass sampler.
+"""Colossus halo mass sampler.
 
-This module facilitates the sampling of halos from `Colossus <http://www.benediktdiemer.com/code/colossus/>`_.
+This module facilitates the sampling of halos from
+`Colossus <http://www.benediktdiemer.com/code/colossus/>`_.
 
-'''
+Models
+======
+.. autosummary::
+    :nosignatures:
+    :toctree: ../api/
+
+    colossus_mass_sampler
+    
+"""
+
 import numpy as np
 import colossus as colossus
 from colossus.lss import mass_function
 from scipy import integrate
 
-def colossus_mass_sampler(redshift, model, mdef, m_min, m_max, cosmology, sigma8, ns, size=None, resolution=1000):
-    """
-        This function generate a sample of halos with halo mass from Despali+16 mass function.
-       
+__all__ = [
+        'colossus_mass_sampler',
+           ]
+
+
+def colossus_mass_sampler(redshift, model, mdef, m_min, m_max,
+                          cosmology, sigma8, ns, size=None, resolution=1000):
+    """Colossus halo mass sampler.
+        This function generate a sample of halos from a mass function which is available in colossus.
+
     Parameters
     -----------
     redshift : float
@@ -27,7 +43,8 @@ def colossus_mass_sampler(redshift, model, mdef, m_min, m_max, cosmology, sigma8
     cosmology : astropy.cosmology.Cosmology
         Astropy cosmology object
     sigma8 : float
-        Cosmology parameter, amplitude of the (linear) power spectrum on the scale of :math:`8 h-1 Mpc`.
+        Cosmology parameter, amplitude of the (linear) power spectrum on the
+        scale of :math:`8 h-1 Mpc`.
     ns : float
         Cosmology parameter, spectral index of scalar perturbation power spectrum.
     size : int, optional
@@ -38,18 +55,20 @@ def colossus_mass_sampler(redshift, model, mdef, m_min, m_max, cosmology, sigma8
     --------
     sample : (size,) array_like
         Samples drawn from the mass function, in units of solar masses.
-        
-    References
+
+        References
     -----------
     .. [1] Diemer et al. (2018) doi 10.3847/1538-4365/aaee8c
     """
-    cosmo = colossus.cosmology.cosmology.fromAstropy(cosmology, sigma8 = sigma8, ns = ns, name = 'my_cosmo')
+    cosmo = colossus.cosmology.cosmology.fromAstropy(cosmology, sigma8=sigma8,
+                                                     ns=ns, name='my_cosmo')
     h0 = cosmo.h
-    m_h0 = np.logspace(np.log10(m_min*h0), np.log10(m_max*h0), resolution) # unit: Msun/h
-    dndm = mass_function.massFunction(m_h0, redshift, mdef = mdef, model = model,q_out = 'dndlnM',q_in='M')/m_h0
+    m_h0 = np.logspace(np.log10(m_min*h0), np.log10(m_max*h0), resolution)  # unit: Msun/h
+    dndm = mass_function.massFunction(m_h0, redshift, mdef=mdef, model=model,
+                                      q_out ='dndlnM', q_in='M')/m_h0
     m = m_h0/h0
     CDF = integrate.cumtrapz(dndm, (m), initial=0)
     CDF = CDF / CDF[-1]
-    n_uniform = np.random.uniform(size = size)
+    n_uniform = np.random.uniform(size=size)
     masssample = np.interp(n_uniform, CDF, m)
     return masssample
