+++ setup.cfg
@@ -34,6 +34,7 @@ test =
     speclite>=0.11
 all =
     speclite>=0.11
+    colossus
 docs =
     sphinx-astropy
     matplotlib
