### 2021-03-04 760b5f2
Consistent use of wavenumber, precise c, corrected tests
