### 2021-03-05 b1dfc93
Update setup.cfg
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
