+++ setup.cfg
@@ -38,6 +38,7 @@ all =
 docs =
     sphinx-astropy
     matplotlib
+    colossus
     speclite>=0.11
 
 [options.package_data]
