### 2021-03-05 30d61fe
Update skypy/halos/sampler.py
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
