### 2021-03-05 de8d5dd
add HAS_COLOSSUS variable and more unit tests
