### 2021-03-05 2e45160
BUG: Fixed wavenumber definition and amplitude scaling in Eisenstein & Hu model (#445)
