+++ skypy/power_spectrum/_eisenstein_hu.py
@@ -6,6 +6,7 @@ formula for the matter power spectrum.
 
 from astropy.utils import isiterable
 import numpy as np
+from astropy import constants
 
 
 __all__ = [
@@ -54,8 +55,8 @@ def transfer_with_wiggles(wavenumber, A_s, n_s, cosmology, kwmap=0.02):
     >>> wavenumber = np.logspace(-3, 1, num=5, base=10.0)
     >>> A_s, n_s = 2.1982e-09, 0.969453
     >>> transfer_with_wiggles(wavenumber, A_s, n_s, Planck15, kwmap=0.02)
-    array([9.92144790e-01, 7.78548704e-01, 1.29998169e-01, 4.63863054e-03,
-       8.87918075e-05])
+    array([9.84445487e-01, 6.77428507e-01, 8.21122028e-02, 2.44250638e-03,
+       4.41412308e-05])
 
     References
     ----------
@@ -84,7 +85,6 @@ def transfer_with_wiggles(wavenumber, A_s, n_s, cosmology, kwmap=0.02):
         raise ValueError("Tcmb0 for input cosmology must be non-zero if"
                          "wiggles = True")
 
-    ak = wavenumber * h0
     om0h2 = om0 * h0**2
     ob0h2 = ob0 * h0**2
     f_baryon = ob0 / om0
@@ -126,8 +126,8 @@ def transfer_with_wiggles(wavenumber, A_s, n_s, cosmology, kwmap=0.02):
     beta_b = 0.5 + f_baryon + (3 - 2 * f_baryon) * np.sqrt((17.2 * om0h2)**2
                                                            + 1.0)
 
-    q = ak / (13.41 * k_eq)
-    ks = ak * sound_horizon
+    q = wavenumber / (13.41 * k_eq)
+    ks = wavenumber * sound_horizon
 
     T_c_ln_beta = np.log(np.e + 1.8 * beta_c * q)
     T_c_ln_nobeta = np.log(np.e + 1.8 * q)
@@ -143,11 +143,11 @@ def transfer_with_wiggles(wavenumber, A_s, n_s, cosmology, kwmap=0.02):
         (1 - T_c_f) * f(T_c_ln_beta, T_c_C_alpha)
 
     s_tilde = sound_horizon * (1 + (beta_node / ks)**3)**(-1 / 3)
-    ks_tilde = ak * s_tilde
+    ks_tilde = wavenumber * s_tilde
 
     T_b_T0 = f(T_c_ln_nobeta, T_c_C_noalpha)
     T_b_1 = T_b_T0 / (1 + (ks / 5.2)**2)
-    T_b_2 = alpha_b / (1 + (beta_b / ks)**3) * np.exp(-(ak / k_silk)**1.4)
+    T_b_2 = alpha_b / (1 + (beta_b / ks)**3) * np.exp(-(wavenumber/k_silk)**1.4)
     T_b = np.sinc(ks_tilde / np.pi) * (T_b_1 + T_b_2)
 
     transfer = f_baryon * T_b + (1 - f_baryon) * T_c
@@ -190,8 +190,8 @@ def transfer_no_wiggles(wavenumber, A_s, n_s, cosmology):
     >>> wavenumber = np.logspace(-3, 1, num=5, base=10.0)
     >>> A_s, n_s = 2.1982e-09, 0.969453
     >>> transfer_no_wiggles(wavenumber, A_s, n_s, Planck15)
-    array([9.91959695e-01, 7.84518347e-01, 1.32327555e-01, 4.60773671e-03,
-       8.78447096e-05])
+    array([9.84321214e-01, 6.83446969e-01, 8.15258290e-02, 2.42055117e-03,
+       4.36725897e-05])
[…]
