### 2021-03-05 119b290
Delete examples in growth
