+++ skypy/power_spectrum/_growth.py
@@ -175,7 +175,7 @@ def growth_function_derivative(redshift, cosmology, gamma=6.0/11.0):
     References
     ----------
     .. [1] E. V. Linder, Phys. Rev. D 72, 043529 (2005)
-    
+
     '''
     z = redshift
 
