### 2021-03-05 20be972
Delete examples Eisenstein & Hu
