### 2021-03-05 5254367
Delete examples in halofit
