### 2021-03-08 f8a49fa
Update skypy/halos/sampler.py
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
