### 2021-03-08 01c5ec9
Update skypy/halos/sampler.py
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
