### 2021-03-09 5f6a144
Implement sampler in _colossus.py and add to mass submodule
