### 2021-03-09 4e2f28a
Clean up cosmology interface
