+++ skypy/halos/tests/test_colossus.py
@@ -9,7 +9,7 @@ from skypy.halos._colossus import HAS_COLOSSUS
 @pytest.mark.flaky
 def test_colossus_mass_sampler():
     from astropy.cosmology import WMAP9
-    import colossus as colossus
+    from colossus.cosmology.cosmology import fromAstropy
     from colossus.lss import mass_function
     from skypy.halos.mass import colossus_mass_sampler
     m_min, m_max, size = 1e+12, 1e+16, 100
@@ -20,9 +20,8 @@ def test_colossus_mass_sampler():
     assert np.all(sample >= m_min)
     assert np.all(sample <= m_max)
     assert np.shape(sample) == (size,)
-    cosmo = colossus.cosmology.cosmology.fromAstropy(WMAP9, sigma8=0.8200,
-                                                     ns=0.9608, name='my_cosmo')
-    h0 = cosmo.h
+    fromAstropy(WMAP9, sigma8=0.8200, ns=0.9608)
+    h0 = WMAP9.h
     m_h0 = np.logspace(np.log10(1e+12*h0), np.log10(1e+16*h0), 1000)
     dndm = mass_function.massFunction(m_h0, 0.1, mdef='500c', model='despali16',
                                       q_out='dndlnM', q_in='M')/m_h0
