### 2021-03-09 24e0ce3
Update skypy/halos/_colossus.py
