### 2021-03-09 8cbb803
Update skypy/halos/_colossus.py
