### 2021-03-09 skypyproject/skypy#444 @rrjbca
@rmorgan10 thanks for the suggestion. We have a few comprehensive example config files, but agree we don't have any simpler examples to teach the syntax and explain how it works so this is important. @Lucia-Fonseca here is the list of syntax features I believe we need to cover:

- [ ] Define a variable
- [ ] Call a function
- [ ] Reference a variable
- [ ] Define the cosmology
- [ ] Define parameters
- [ ] Create a table
- [ ] Add a column
- [ ] Multi-column assignment
- [ ] table.init and table.complete dependencies
