### 2021-03-09 1aa09dc
ENH: Add a mass sampler for the colossus halo mass function (#423)
Co-authored-by: Richard R <58728519+rrjbca@users.noreply.github.com>
