+++ setup.cfg
@@ -34,9 +34,11 @@ test =
     speclite>=0.11
 all =
     speclite>=0.11
+    colossus
 docs =
     sphinx-astropy
     matplotlib
+    colossus
     speclite>=0.11
 
 [options.package_data]
