### 2021-03-11 d0c1115
Implement colossus_mf_redshift and colossus_mf
