### 2021-03-11 4292eb7
Unit tests for colossus_mf_redshift and colossus_mf
