### 2021-03-11 0886c6a
Refactor halo submodules and documentation to accommodate new functions
