### 2021-03-12 bc59000
Update skypy/halos/_colossus.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
