### 2021-03-12 4d7182f
Remove skypy.halos.redshift submodule
