### 2021-03-14 skypyproject/skypy#443 @cescalara
Hi, I agree with @rmorgan10. For me, there is a gap between the high-level software description in the docs/paper and the examples which jump into to specific applications. I think the suggested text in the docs/paper could also benefit form a discussion of the following points:

* What are some example scientific questions that SkyPy tools can help address with its current tools?
* As noted above, it looks as though the developers plan to add more functionality. What is the scope of the package? Is the focus on simulation of observed galaxy populations, or populations of other astrophysical objects too? It seems SkyPy is already popular and such a statement could help potential contributors understand what you could be looking for in the future and where the project is going.
