### 2021-03-14 #3056 @cescalara
Hi @rrjbca, I am impressed with SkyPy. The code was easy to install, test and get working with quickly. The code itself is also carefully documented. In general, I agree with the recommendations and issues raised by @rmorgan10, and have added a comment in [#433](https://github.com/skypyproject/skypy/issues/443) with my thoughts. I also add some further comments/questions below.

* Can you clarify the contributions of co-authors who have minor/no contributions to the code?
* As I understand, at the moment SkyPy focusses on simulating observed galaxy properties as seen by different detectors. Are there plans to include the simulation of undetected objects and stochastic selection effects?
* All the examples run smoothly, but it would be nice to have some optional logging to show the user what the code is up to. This would be helpful for users designing and debugging their own pipelines.
* It seems that there is a small typo or bug on [this page](https://skypy.readthedocs.io/en/latest/examples/galaxies/plot_photometry.html#sphx-glr-examples-galaxies-plot-photometry-py) in the docs. When I run `$ skypy examples/galaxies/sdss_photometry.yml sdss_photometry.fits` I get the following error `skypy: error: unrecognized arguments: sdss_photometry.fits`.
* Can you clarify the relationship to other packages with similar functionality/goals? I am aware of the following python packages which seem to have some overlap in functionality/scope (correct me if I am wrong).
  * https://github.com/GalSim-developers/GalSim
  * https://github.com/astropy/halotools

I hope these comments are useful and am happy to discuss on any of the points. Thank you all for your work on SkyPy.
