### 2021-03-17 529b17c
Logging for skypy command line script and Pipeline class
