### 2021-03-18 4d06a4e
Test logging by skypy command line script
