### 2021-03-18 6dbe375
Add logging instruction to documentation
