+++ docs/pipeline/index.rst
@@ -21,6 +21,9 @@ fits files:
 
     $ skypy examples/galaxies/sdss_photometry.yml sdss_photometry.fits
 
+To view the progress of the pipeline as it runs you can enable logging using the
+`--logging INFO` argument.
+
 Config files are written in YAML format and read using the
 `~skypy.pipeline.load_skypy_yaml` funciton. Each entry in the config specifices
 an arbitrary variable, but there are also some particular fields that SkyPy uses:
