### 2021-03-22 8af5853
Switch to verbose/quiet command line flags
