+++ skypy/pipeline/scripts/skypy.py
@@ -15,8 +15,10 @@ def main(args=None):
     parser.add_argument('output', help='Output file name')
     parser.add_argument('-o', '--overwrite', action='store_true',
                         help='Whether to overwrite existing files')
-    parser.add_argument('-l', '--logging', default='NOTSET', help='logging level',
-                        choices=[n for n in logging._nameToLevel if n != 'NOTSET'])
+    parser.add_argument("-v", "--verbose", action="count", default=0,
+                        help="Increase logging verbosity")
+    parser.add_argument("-q", "--quiet", action="count", default=0,
+                        help="Decrease logging verbosity")
 
     # get system args if none passed
     if args is None:
@@ -25,12 +27,14 @@ def main(args=None):
     args = parser.parse_args(args or ['--help'])
 
     # Setup skypy logger
+    default_level = logging._nameToLevel['WARNING']
+    logging_level = default_level + 10 * (args.quiet - args.verbose)
     formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(name)s: %(message)s')
     stream_handler = logging.StreamHandler()
-    stream_handler.setLevel(args.logging)
+    stream_handler.setLevel(logging_level)
     stream_handler.setFormatter(formatter)
     logger = logging.getLogger('skypy')
-    logger.setLevel(args.logging)
+    logger.setLevel(logging_level)
     logger.addHandler(stream_handler)
 
     try:
