### 2021-03-22 8233eca
Update unit tests for verbose/quiet command line flags
