### 2021-03-22 f719c5d
Log pipeline jobs only if they are not skipped
