### 2021-03-22 ede99e6
Log writing output file
