+++ skypy/pipeline/_items.py
@@ -1,6 +1,7 @@
 '''item types in the pipeline'''
 
 from collections.abc import Sequence, Mapping
+from . import log
 import inspect
 
 
@@ -88,4 +89,5 @@ class Call(Item):
         '''execute the call in the given pipeline'''
         args = pipeline.evaluate(self.args)
         kwargs = pipeline.evaluate(self.kwargs)
+        log.info(f"{self.function.__name__}")
         return self.function(*args, **kwargs)
