### 2021-03-22 1babb99
Fix unit test for default table and flake8
