### 2021-03-22 54054ce
Unit test logging cosmology
