### 2021-03-22 3a56b8c
Modify test_config.yml to use cosmology keyword
