### 2021-03-23 f390eda
Add section for creating config files to docs
