### 2021-03-23 273ccec
Add guide for variables
