+++ docs/how_to_config_file.rst
@@ -12,17 +12,41 @@ functions and pipelines!
 YAML syntax
 -----------
 YAML_ is a file format designed to be readable by both computers and humans.
-This guide introduces the main features of YAML relevant when writing
-a configuration file to use with ``SkyPy``.
 Fundamentally, a file written in YAML consists of a set of key-value pairs.
 Each pair is written as ``key: value``, where whitespace after the ``:`` is optional.
 
+This guide introduces the main syntax of YAML relevant when writing
+a configuration file to use with ``SkyPy``. Essentially, it would start with
+definitions of individual variables at the top, followed by the tables to produce,
+and, within the table entries, the features of the objects to simulate are included.
 
 
 Variables
 ^^^^^^^^^
-* Define a variable
-* Reference a variable
+* `Define a variable`: a variable is defined as a key-value pair at the top of the file. You can define any type of variable:
+
+  .. code:: yaml
+
+      counter: 100 # An integer
+      miles: 1000.0 # A floating point
+      name: "Joy" # A string
+      mylist: [ 'abc', 789, 2.0e3 ] # A list
+
+
+* `Reference a variable`: variables can be referenced by their full name tagged with a dollar sign ``$``.
+  For example, if you previously defined a list of countries:
+
+  .. code:: yaml
+
+      country_list: [ 'UK', 'Spain', 'Zambia', 'Chile', 'Japan' ]
+
+  You would reference the variable;
+
+  .. code:: yaml
+
+      countries: $country_list
+
+
 
 Functions
 ^^^^^^^^^
