### 2021-03-23 832c6f5
Add guide for functions
