### 2021-03-23 bc2945f
Add guide for cosmology
