### 2021-03-23 2e65b25
Change example for functions
