### 2021-03-23 57ea54d
Add example for .ini and depends
