### 2021-03-29 faf7085
PowerSpectrum base classes
