### 2021-03-29 390f12d
EisensteinHu class
