### 2021-03-29 cd0550b
Add power spectrum instances to submodule
