+++ skypy/power_spectrum/_camb.py
@@ -5,6 +5,7 @@ from ._base import TabulatedPowerSpectrum
 
 
 __all__ = [
+    'CAMB',
     'camb',
 ]
 
