### 2021-03-29 cbe36de
Fix handling of scalar and unsorted arguments
