### 2021-03-29 4acdb4a
Reduce degree of interploation spline for sparse data
