### 2021-03-29 8a4cb0e
Fix name of halo function
