### 2021-03-29 6be3489
Update CAMB tests to use power spectrum interpolator
