+++ skypy/power_spectrum/tests/test_camb.py
@@ -22,18 +22,23 @@ def test_camb():
     '''
     Test a default astropy cosmology
     '''
-    from skypy.power_spectrum import camb
+    from skypy.power_spectrum import CAMB
+
+    # Setup CAMB interpolator
+    k_max, z_grid = 10, np.array([0, 1])
+    A_s, n_s = 2.e-9, 0.965
+    ps = CAMB(k_max, z_grid, Planck15, A_s, n_s)
 
     # test shape and compare with the mocked power spectrum
     redshift = [0.0, 1.0]
     wavenumber = np.logspace(-4.0, np.log10(2.0), 200)
-    pzk = camb(wavenumber, redshift, Planck15, 2.e-9, 0.965)
+    pzk = ps(wavenumber, redshift)
     assert pzk.shape == (len(redshift), len(wavenumber))
     assert allclose(pzk, test_pzk, rtol=1.e-4)
 
     # also check redshifts are ordered correctly
     redshift = [1.0, 0.0]
-    pzk = camb(wavenumber, redshift, Planck15, 2.e-9, 0.965)
+    pzk = ps(wavenumber, redshift)
     assert pzk.shape == (len(redshift), len(wavenumber))
     assert allclose(pzk, test_pzk[np.argsort(redshift), :], rtol=1.e-4)
 
@@ -44,10 +49,16 @@ def test_camb_redshift_zero():
     Regression test for #438
     Test that camb runs succesfully with redshift=0.0
     """
-    from skypy.power_spectrum import camb
+    from skypy.power_spectrum import CAMB
+
+    # Setup CAMB interpolator
+    k_max, z_grid = 10, np.array([0, 1])
+    A_s, n_s = 2.e-9, 0.965
+    ps = CAMB(k_max, z_grid, Planck15, A_s, n_s)
+
     redshift = 0.0
     wavenumber = np.logspace(-4.0, np.log10(2.0), 200)
-    pzk = camb(wavenumber, redshift, Planck15, 2.e-9, 0.965)
+    pzk = ps(wavenumber, redshift)
     assert allclose(pzk, test_pzk[0], rtol=1.e-4)
 
 
