### 2021-03-29 38bf2e7
Change style for walkthrough example
