### 2021-03-30 df52c7d
Test EisensteinHu class and fix bug when using scalar redshift argument
