### 2021-03-31 7cba633
Add main functionalities page in doc
