+++ docs/main_functionalities.rst
@@ -1,0 +1,5 @@
+####################
+Main functionalities
+####################
+
+This page outlines the main functionalities in the `SkyPy` library.
