+++ docs/main_functionalities.rst
@@ -3,3 +3,19 @@ Main functionalities
 ####################
 
 This page outlines the main functionalities in the `SkyPy` library.
+
+Galaxies
+--------
+- `Galaxy Morphology Distributions`_
+- `Galaxy Redshift Distributions`_
+
+
+.. _Galaxy Morphology Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.morphology
+.. _Galaxy Redshift Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.redshift
+
+Utilities
+---------
+
+
+Pipeline
+--------
