### 2021-03-31 38bd887
Add luminosity, spectra and stellar mass
