### 2021-03-31 a86caff
Add pipeline functionality
