+++ docs/main_functionalities.rst
@@ -30,5 +30,11 @@ Utils
 .. _Random Distributions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.random
 .. _Special Functions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.special
 
+
 Pipeline
 --------
+- `Pipeline`_:
+    - Running `SkyPy` from the command line
+    - Using a pipeline from other code
+
+.. _Pipeline: https://skypy.readthedocs.io/en/latest/pipeline/index.html
