### 2021-04-06 44b3c8a
Fix SkyPy example
