### 2021-04-13 f3b542b
Rename config file section on main doc page
