### 2021-04-13 4afdec0
Edit wording in example section as it now comes first
