### 2021-04-22 e71cfde
Commen on running pipeline from command line
