### 2021-04-22 9732f18
Adding link of speclite filters to documentation.
