### 2021-04-22 skypyproject/skypy#444 @philipp128
> Furthermore, a specific request I have for this sort of guide is to list all the filters that can be used. I was unable to find what the other options instead of "sdss2010-*" were in the galaxy photometry example. If you could point out where this documentation is (if I just missed it) that would also work for me.

PR #457 addresses  this.
