### 2021-04-22 7058478
Move variable referencing to the functions section
