### 2021-04-22 41543b8
Fix functions example
