### 2021-04-22 2b1114f
added get in touch, expanded summary
