### 2021-04-22 18eafdc
added link to main_functionalities page, typos
