### 2021-04-22 1281106
ENH: Logging for skypy command line script and Pipeline class (#453)
