+++ skypy/pipeline/tests/data/test_config.yml
@@ -3,7 +3,7 @@ test_float: 1.0
 test_str: hello world
 test_func: !numpy.random.uniform
 test_func_with_arg: !len 'hello world'
-test_cosmology: !astropy.cosmology.FlatLambdaCDM
+cosmology: !astropy.cosmology.FlatLambdaCDM
   H0: 67.74
   Om0: 0.3075
 tables:
