### 2021-04-22 skypyproject/skypy#444 @rmorgan10
> PR #457 addresses this.

This [additional documentation](https://speclite.readthedocs.io/en/latest/filters.html) looks great! Thanks for adding it
