+++ README.rst
@@ -24,6 +24,7 @@ visit the SkyPy Collaboration website: `http://skypyproject.org`_
 .. _Pipelines: https://skypy.readthedocs.io/en/latest/pipeline/index.html
 .. _full list of features: https://skypy.readthedocs.io/en/latest/main_functionalities.html
 .. _SkyPy Documentation: https://skypy.readthedocs.io/en/latest/
+.. _http://skypyproject.org: http://skypyproject.org
 
 Citation
 --------
