### 2021-04-22 756e4c8
Fix cosmic voids example variance
