### 2021-04-22 5608612
Fix circular velocity example
