### 2021-04-27 f42f226
Restructure text around table referencing
