### 2021-04-27 c6ccf18
Use numpy functions in abundance matching example
