### 2021-04-27 5bc5a56
Add clone example in cosmology section
