+++ skypy/pipeline/_config.py
@@ -52,36 +52,21 @@ class SkyPyLoader(yaml.SafeLoader):
         return Ref(ref)
 
     def construct_call(self, name, node):
-        if isinstance(node, yaml.ScalarNode):
-            arg = self.construct_scalar(node)
-            args = [arg] if arg != '' else []
-            kwargs = {}
-        elif isinstance(node, yaml.SequenceNode):
-            args = self.construct_sequence(node)
-            kwargs = {}
-        elif isinstance(node, yaml.MappingNode):
-            args = []
-            kwargs = self.construct_mapping(node)
-
-        try:
-            function = import_function(name)
-        except (ModuleNotFoundError, AttributeError) as e:
-            raise ImportError(f'{e}\n{node.start_mark}') from e
-
-        return Call(function, args, kwargs)
-
-    def construct_imported_object(self, node):
-        name = self.construct_scalar(node)
-        if name[:1] == '~':
-            name = name[1:]
-        if not name:
-            raise ValueError(f'empty object\n{node.start_mark}')
         try:
             object = import_function(name)
         except (ModuleNotFoundError, AttributeError) as e:
             raise ImportError(f'{e}\n{node.start_mark}') from e
 
-        return object
+        if isinstance(node, yaml.ScalarNode):
+            return object
+        else:
+            if isinstance(node, yaml.SequenceNode):
+                args = self.construct_sequence(node)
+                kwargs = {}
+            if isinstance(node, yaml.MappingNode):
+                args = []
+                kwargs = self.construct_mapping(node)
+            return Call(object, args, kwargs)
 
     def construct_quantity(self, node):
         value = self.construct_scalar(node)
@@ -95,10 +80,6 @@ SkyPyLoader.add_implicit_resolver('!ref', re.compile(r'\$\w+'), ['$'])
 
 # constructor for generic function calls
 SkyPyLoader.add_multi_constructor('!', SkyPyLoader.construct_call)
-
-# constructor for importing generic objects from external (sub)modules
-SkyPyLoader.add_constructor('!import', SkyPyLoader.construct_imported_object)
-SkyPyLoader.add_implicit_resolver('!import', re.compile(r'\~\w+'), ['~'])
 
 # constructor for quantities
 SkyPyLoader.add_constructor('!quantity', SkyPyLoader.construct_quantity)
