### 2021-04-27 f9940ed
DOC: Describe speclite filters in documentation (#457)
