+++ docs/main_functionalities.rst
@@ -8,25 +8,25 @@ Galaxies
 --------
 
 - `Luminosity Distributions`_
-- `Morphology Distributions`_
+- `Morphological Distributions`_
 - `Redshift Distributions`_
-- `Spectrum Distributions`_
+- `Spectral Energy Distribution Modelling`_
 - `Stellar Mass Distributions`_
 
 .. _Luminosity Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.luminosity
-.. _Morphology Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.morphology
+.. _Morphological Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.morphology
 .. _Redshift Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.redshift
-.. _Spectrum Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.spectrum
+.. _Spectral Energy Distribution Modelling: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.spectrum
 .. _Stellar Mass Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.stellar_mass
 
 Utils
 -----
 
-- `Photometry`_ : spectral energy distributions and photometric properties.
+- `Photometric Calculations`_
 - `Random Distributions`_
 - `Special Functions`_
 
-.. _Photometry: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.photometry
+.. _Photometric Calculations: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.photometry
 .. _Random Distributions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.random
 .. _Special Functions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.special
 
