### 2021-04-27 b6e2ed8
Delete non astrophysical features
