+++ docs/main_functionalities.rst
@@ -18,23 +18,3 @@ Galaxies
 .. _Redshift Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.redshift
 .. _Spectral Energy Distribution Modelling: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.spectrum
 .. _Stellar Mass Distributions: https://skypy.readthedocs.io/en/latest/galaxies.html#module-skypy.galaxies.stellar_mass
-
-Utils
------
-
-- `Photometric Calculations`_
-- `Random Distributions`_
-- `Special Functions`_
-
-.. _Photometric Calculations: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.photometry
-.. _Random Distributions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.random
-.. _Special Functions: https://skypy.readthedocs.io/en/latest/utils/index.html#module-skypy.utils.special
-
-
-Pipeline
---------
-- `Pipeline`_:
-    - Running `SkyPy` from the command line
-    - Using a pipeline from other code
-
-.. _Pipeline: https://skypy.readthedocs.io/en/latest/pipeline/index.html
