### 2021-04-27 cf4eaa0
Update funciton syntax in unit tests, add unit test for object import
