### 2021-04-27 37ed1f6
Planck18 not in oldest supported version of astropy, use Planck15 in test instead
