### 2021-04-27 ce1abb2
Raise ValueError if ScalarNode has a non-empty value
