### 2021-04-27 eb43a42
ENH: Config syntax for importing objects (#463)
