### 2021-05-04 51c201b
Revert "strip gravitational_wave module"
This reverts commit a457b5e807ec27a75e945930189572718a4da8f4.
