+++ skypy/gravitational_waves/tests/test_import.py
@@ -1,0 +1,2 @@
+def test_import():
+    import skypy.gravitational_waves
