+++ skypy/gravitational_waves/__init__.py
@@ -1,0 +1,16 @@
+"""
+This module contains methods that model the properties of gravitational wave
+populations.
+
+Merger Rates
+================
+
+.. autosummary::
+   :nosignatures:
+   :toctree: ../api/
+
+   b_band_merger_rate
+
+"""
+
+from .merger_rate import *  # noqa F401,F403
