+++ docs/gravitational_waves/index.rst
@@ -1,0 +1,5 @@
+*************************************************
+Gravitational Waves (`skypy.gravitational_waves`)
+*************************************************
+
+.. automodule:: skypy.gravitational_waves
