+++ skypy/neutral_hydrogen/tests/test_import.py
@@ -1,2 +1,2 @@
 def test_import():
-    import skypy.neutral_hydrogen
+    import skypy.neutral_hydrogen  # noqa
