### 2021-05-12 eeec851
Fix typ for Planck13 cosmology
