+++ docs/configuration_files.rst
@@ -336,7 +336,7 @@ Import objects
 
   .. code:: yaml
 
-    cosmo1: !astropy.cosmology.Planck13 # import the Planck18 object and bind to cosmo1
+    cosmo1: !astropy.cosmology.Planck13 # import the Planck13 object and bind to cosmo1
     cosmo2: !astropy.cosmology.default_cosmology.get [] # call default_cosmology.get() and bind to cosmo2
 
 
