### 2021-05-12 d7d4f0e
Delete how-not-to-do example
