### 2021-05-12 94c7160
Fix plot rendering
