### 2021-05-13 51bf442
Regression test for bug #464
