### 2021-05-13 c3f9c27
Fix typo argument cosmology
