+++ docs/luminosity.yml
@@ -1,4 +1,4 @@
-cosmology: !astropy.cosmology.default_cosmology.get
+cosmology: !astropy.cosmology.default_cosmology.get []
 z_range: !numpy.linspace [0, 2, 21]
 M_star: !astropy.modeling.models.Linear1D [-0.9, -20.4]
 phi_star: !astropy.modeling.models.Exponential1D [3e-3, -9.7]
