### 2021-05-14 6e52bbb
Update skypy/pipeline/tests/test_skypy.py
