### 2021-05-14 f75a17e
Add assert statement to regression test
