### 2021-05-14 149d3aa
Remove config file for old regression test
