### 2021-05-14 4d121ee
Fix argument default cosmology
