### 2021-05-14 b6d9dd2
Change to a smaller sky area
