### 2021-05-14 c9857e9
Decrease sky area in the luminosity config file
