### 2021-05-14 a75a57d
Fix telescope example
