### 2021-05-14 e674976
Fix telescope example 2
