### 2021-05-14 88bd0bc
Fix telescope example to have coumns with same length
