### 2021-05-14 abd98de
Fix RuntimeWarning for the circular velocities examples
