### 2021-05-14 3459113
Fix indentation problem in luminosity pipeline
