### 2021-05-14 1d2f201
Add cosmology bullet point in the functions section
