### 2021-05-14 4e163a3
Add temporary example for the cosmology feature
