### 2021-05-20 e0404c6
Add .depends under functions and include example
