### 2021-05-20 7164141
Add table initialisation section and example
