+++ docs/configuration_files.rst
@@ -6,8 +6,8 @@ This page outlines how to construct configuration files to run your own routines
 with `~skypy.pipeline.Pipeline`.
 
 `SkyPy` is an astrophysical simulation pipeline tool that allows to define any
-arbitrary workflow and store data in table format. You may use `SkyPy` `~skypy.pipeline.Pipeline`.
-to call any function --your own, from any compatible external software or from the `SkyPy library`.
+arbitrary workflow and store data in table format. You may use `SkyPy` `~skypy.pipeline.Pipeline`
+to call any function --your own implementation, from any compatible external software or from the `SkyPy library`.
 Then `SkyPy` deals with the data dependencies and provides a library of functions to be used with it.
 
 These guidelines start with an example using one of the `SkyPy` functions, and it follows
