### 2021-06-08 11d4a4f
Fix .init example
