### 2021-06-08 d1abe32
Fix .depends example by adding cosmology
