+++ docs/configuration_files.rst
@@ -109,6 +109,9 @@ YAML syntax
 YAML_ is a file format designed to be readable by both computers and humans.
 Fundamentally, a file written in YAML consists of a set of key-value pairs.
 Each pair is written as ``key: value``, where whitespace after the ``:`` is optional.
+The hash character (#) denotes the start of a comment and all further text on that
+line is ignored by the parser.
+
 
 This guide introduces the main syntax of YAML relevant when writing
 a configuration file to use with ``SkyPy``. Essentially, it begins with
