### 2021-06-15 034814b
DOC: Remove docstring examples (#429)
