### 2021-06-15 030e769
DOC: Remove Docstring examples in power_spectrum (#447)
