+++ docs/conf.py
@@ -30,8 +30,6 @@ import sys
 import datetime
 from importlib import import_module
 
-import skypy
-
 try:
     from sphinx_astropy.conf.v1 import *  # noqa
 except ImportError:
@@ -64,11 +62,7 @@ exclude_patterns.append('_templates')
 # This is added to the end of RST files - a good place to put substitutions to
 # be used globally.
 rst_epilog += """
-.. |minimum_python_version| replace:: {0.__minimum_python_version__}
-.. |minimum_numpy_version| replace:: {0.__minimum_numpy_version__}
-.. |minimum_scipy_version| replace:: {0.__minimum_scipy_version__}
-.. |minimum_astropy_version| replace:: {0.__minimum_astropy_version__}
-""".format(skypy)
+"""
 
 # -- Project information ------------------------------------------------------
 
