+++ docs/install.rst
@@ -1,202 +1,1 @@
-************
-Installation
-************
 
-.. _astropy-main-req:
-
-Requirements
-============
-
-``SkyPy`` has the following strict requirements:
-
-- `Python <https://www.python.org/>`_ |minimum_python_version| or later
-
-- `Numpy`_ |minimum_numpy_version| or later
-
-- `scipy`_ |minimum_scipy_version| or later: To power a variety of features
-  in several modules.
-
-- `astropy`_ |minimum_astropy_version| or later
-
-- `networkx <https://networkx.github.io/>`_
-
-- `PyYAML <https://pyyaml.org>`_
-
-``SkyPy`` also depends on other packages for optional features:
-
-- `matplotlib <https://matplotlib.org/>`_ 2.0 or later: To provide plotting
-  functionality that `astropy.visualization` enhances.
-
-- `pytest-astropy <https://github.com/astropy/pytest-astropy>`_
-
-- `sphinx-astropy <https://github.com/astropy/sphinx-astropy>`_
-
-- `camb <<https://github.com/astropy/sphinx-astropy>`_
-.. - `setuptools <https://setuptools.readthedocs.io>`_: Used for discovery of
-  entry points which are used to insert fitters into `astropy.modeling.fitting`.
-
-
-The following packages can optionally be used when testing:
-
-- `tox <https://tox.readthedocs.io/en/latest/>`_: Used to automate testing
-  and documentation builds.
-
-
-Installing ``SkyPy``
-======================
-
-If you are new to Python and/or do not have familiarity with `Python virtual
-environments <https://docs.python.org/3/tutorial/venv.html>`_, then we recommend
-starting by installing the `Anaconda Distribution
-<https://www.anaconda.com/distribution/>`_. This works on all platforms (linux,
-Mac, Windows) and installs a full-featured scientific Python in a user directory
-without requiring root permissions.
-
-Using pip
----------
-
-.. warning::
-
-    Users of the Anaconda Python distribution should follow the instructions
-    for :ref:`anaconda_install`.
-
-To install ``SkyPy`` with `pip <https://pip.pypa.io>`__, run::
-
-    pip install skypy
-
-If you want to make sure none of your existing dependencies get upgraded, you
-can also do::
-
-    pip install skypy --no-deps
-
-..On the other hand, if you want to install ``SkyPy`` along with all of the
-.. available optional dependencies, you can do::
-
-..    pip install astropy[all]
-
-In most cases, this will install a pre-compiled version (called a *wheel*) of
-SkyPy, but if you are using a very recent version of Python, if a new version
-of SkyPy has just been released, or if you are building SkyPy for a platform
-that is not common, SkyPy will be installed from a source file. Note that in
-this case you will need a C compiler (e.g., ``gcc`` or ``clang``) to be installed
-(see `Building from source`_ below) for the installation to succeed.
-
-If you get a ``PermissionError`` this means that you do not have the required
-administrative access to install new packages to your Python installation. In
-this case you may consider using the ``--user`` option to install the package
-into your home directory. You can read more about how to do this in the `pip
-documentation <https://pip.pypa.io/en/stable/user_guide/#user-installs>`_.
-
-Do **not** install ``SkyPy`` or other third-party packages using ``sudo``
-unless you are fully aware of the risks.
-
-.. .. _anaconda_install:
-
-.. Using Conda
-.. -----------
-
-.. To install ``SkyPy`` using conda run::
-
-..     conda install skypy
-
-.. ``SkyPy`` is installed by default with the `Anaconda Distribution
-.. <https://www.anaconda.com/distribution/>`_. To update to the latest version run::
-
-..     conda update skypy
-
-.. There may be a delay of a day or two between when a new version of ``SkyPy``
[…]
