### 2021-06-22 c34f884
Apply suggestions from code review
