### 2021-06-22 15e1ded
DOC: Expanded landing page documentation (#228)
