+++ paper.md
@@ -11,7 +11,7 @@ authors:
     affiliation: 1
   - name: Lucia F. de la Bella
     orcid: 0000-0002-1064-3400
-    affiliation: 2
+    affiliation: "1, 2"
   - name: Simon Birrer
     orcid: 0000-0003-3195-5507
     affiliation: 3
@@ -26,7 +26,7 @@ authors:
     affiliation: 4
   - name: Ian Harrison
     orcid: 0000-0002-4437-0770
-    affiliation: 2
+    affiliation: "5, 2"
   - name: Ian W. Harry
     orcid: 0000-0002-5304-9372
     affiliation: 1
@@ -38,28 +38,28 @@ authors:
     affiliation: 1
   - name: Brian Nord
     orcid: 0000-0001-6706-8972
-    affiliation: "5, 6, 7"
+    affiliation: "6, 7, 8"
   - name: Laura K. Nuttall
     orcid: 0000-0002-8599-8791
     affiliation: 1
-  - name: Richard P. Rollins^[richard.rollins\@manchester.ac.uk]
+  - name: Richard P. Rollins^[richard.rollins\@ed.ac.uk]
     orcid: 0000-0003-1291-1023
-    affiliation: 2
+    affiliation: "9, 2"
   - name: Philipp Sudek
     orcid: 0000-0001-8685-2308
     affiliation: 1
   - name: Sut-Ieng Tam
     orcid: 0000-0002-6724-833X
-    affiliation: 8
+    affiliation: 10
   - name: Nicolas Tessore
     orcid: 0000-0002-9696-7931
-    affiliation: 9
+    affiliation: 11
   - name: Arthur E. Tolley
     orcid: 0000-0001-9841-943X
     affiliation: 1
   - name: Keiichi Umetsu
     orcid: 0000-0002-7196-4822
-    affiliation: 8
+    affiliation: 10
   - name: Andrew R. Williamson
     orcid: 0000-0002-7627-8688
     affiliation: 1
@@ -75,31 +75,38 @@ affiliations:
     index: 3
   - name: Institute of Physics, Laboratory of Astrophysics, Ecole Polytechnique Fédérale de Lausanne
     index: 4
-  - name: Fermi National Accelerator Laboratory
+  - name: Department of Physics, University of Oxford
     index: 5
-  - name: Kavli Institute for Cosmological Physics, University of Chicago
+  - name: Fermi National Accelerator Laboratory
     index: 6
-  - name: Department of Astronomy and Astrophysics, University of Chicago
+  - name: Kavli Institute for Cosmological Physics, University of Chicago
     index: 7
-  - name: Institute of Astronomy and Astrophysics, Academia Sinica
+  - name: Department of Astronomy and Astrophysics, University of Chicago
     index: 8
-  - name: Department of Physics and Astronomy, University College London
+  - name: Institute for Astronomy, University of Edinburgh
     index: 9
-date: 15 February 2021
+  - name: Institute of Astronomy and Astrophysics, Academia Sinica
+    index: 10
+  - name: Department of Physics and Astronomy, University College London
+    index: 11
+
+date: 30 June 2021
 bibliography: paper.bib
 
 ---
 
 # Summary
 
[…]
