### 2021-06-30 #3056 @rrjbca
Thank you all again for your kind and constructive comments. As a result we have now revised the manuscript and made a number of improvements to the code and documentation. We address the specific points raised in your reviews below:

@rmorgan10
- In [PR #456](https://github.com/skypyproject/skypy/pull/456) we added a [Feature List](https://skypy.readthedocs.io/en/latest/feature_list.html) to the documentation outlining the galaxy properties that we implement models for in the library. We also added a second paragraph in the “Summary” section of the manuscript to include the high-level description of the package’s functionality.
- In [PR #454](https://github.com/skypyproject/skypy/pull/454) we added a [Configuration Files](https://skypy.readthedocs.io/en/latest/configuration_files.html) section to the documentation demonstrating and explaining how to construct configuration files.
- In [PR #457](https://github.com/skypyproject/skypy/pull/457) we clarify the documentation for our [Galaxy Spectrum](https://skypy.readthedocs.io/en/latest/galaxies.html#spectrum) and [Photometry](https://skypy.readthedocs.io/en/latest/utils/index.html#photometry-skypy-utils-photometry) modules stating that we use [speclite](https://speclite.readthedocs.io/en/latest/) for our photometric calculations and that users should refer to their documentation for the list of available filters and instructions to use your own filters.

@cescalara
- The list of co-authors on the paper includes all members of the [SkyPy Collaboration](https://skypyproject.org/members/), each of whom has directly contributed to the preparation of the manuscript. Those with no commits to the repository have still made important contributions to the software in areas such as planning, design, management and review which we believe are important to recognise.
- Currently SkyPy can simulate galaxies down to a given apparent magnitude or stellar-mass limit. It is up to the user how they wish to interpret the output of these functions. For example, a user could naively use SkyPy to simulate an “observed” noise-free magnitude-limited sample. They could also simulate a deeper sample and apply their own empirical selection, or use the sample to generate their own image simulations. These additional processing steps are not explicitly implemented within SkyPy, but the modular design allows users to integrate external software implementing these process into a SkyPy pipeline.
- Thank you for the suggestion, this has now been implemented in [PR #453](https://github.com/skypyproject/skypy/pull/453) and can be run in the development version using the command line flag `--verbose` / `-v` to log function calls and file I/O.
- It looks like you have installed the most recent stable version 0.4 but are following instructions in the documentation for the development branch. Could you please try again following the instructions in the [v0.4 documentation](https://skypy.readthedocs.io/en/stable/examples/galaxies/plot_photometry.html#kcorrect-spectral-templates)? Specifically you should run `skypy sdss_photometry.yml --format fits`
- The v0.4 SkyPy release currently does not implement either image simulations or models for the galaxy-halo connection so there is not currently any notable overlap. However, as we develop new features we will continue to reuse existing high-quality software where possible, preferring to write interfaces between the SkyPy Pipeline infrastructure and libraries such as GalSim and Halotools. We added an extra line at the end of the Summary section in the manuscript to clarify this point.

@arfon

- In addressing the reviewers comments we have done significant work to improve the software and documentation on our development branch. Would you expect us to make a new release (nominally v0.4.1) incorporating these changes to accompany the paper?
- The following citation macro in markdown `[@Astropy2013; @Astropy2018]` leads to the following citation in the manuscript: “(Astropy Collaboration, 2018, 2013)”. The references are also listed in reverse-chronological order in the bibliography. Is it possible to force these citations to be in chronological order?
