### 2021-07-01 #3056 @arfon
> In addressing the reviewers comments we have done significant work to improve the software and documentation on our development branch. Would you expect us to make a new release (nominally v0.4.1) incorporating these changes to accompany the paper?

Yes please. A new release would be great.

> The references are also listed in reverse-chronological order in the bibliography. Is it possible to force these citations to be in chronological order?

I'm not sure sorry. Once we're a little closer to publishing, I'll download the paper locally and see if there's something I can do to fix this.
