### 2021-07-06 #3056 @cescalara
Hi @rrjbca, everything looks good, I really like the improved docs and new logging functionality. Thank you!

One small thing, now that I understand better the scope of SkyPy, perhaps there is also some relation to [popsynth](https://github.com/grburgess/popsynth), a package that I am using in my work. In particular regarding luminosity function/density evolution simulation and the idea to have a general framework for generative modelling and stats applications. I understand that the focus for applications of SkyPy is different but there seem to be some similarities. Perhaps it could be worth mentioning for the "state of the field" bullet point. If there is really nothing out there that is related, this could also be worth mentioning explicitly!

Otherwise I accept the code and paper as is.
