### 2021-07-09 #3056 @cescalara
Hi again @rrjbca, upon a closer look I did come across something that I'd like to check. I noticed  that when sampling redshift distributions and luminosity functions (e.g. [here](https://github.com/skypyproject/skypy/blob/15e1ded0ea73173666934a3efbc66a2286802465/skypy/galaxies/redshift.py#L208) and [here](https://github.com/skypyproject/skypy/blob/15e1ded0ea73173666934a3efbc66a2286802465/skypy/utils/random.py#L21)) an approximation of the inverse-CDF method is used. In my experience, this is very fragile and can give biased results, with this effect depending on the size of the discretisation used.

I would recommend to replace this with simple rejection sampling for a robust approach that should be fast enough for the envisaged applications. If you prefer not to do this, then I would suggest to _at least_ warn the user and suggest reasonable parameter ranges/resolutions for which the function has been tested, and to what precision.

Another smaller comment, the use of `noise` [here](https://github.com/skypyproject/skypy/blob/15e1ded0ea73173666934a3efbc66a2286802465/skypy/galaxies/redshift.py#L229) is misleading from a pedagogical stats perspective and I would recommend to change this to something like `fixed_N`. It isn't really noise, just whether you are assuming that your observations arise from a poisson process.

I apologise for missing this in my earlier reviews.
