### 2021-07-13 91c6d4c
Add placeholder for ellipticity example
