### 2021-07-13 6481aa1
DOC: Inverse transform sampling accuracy warning
