### 2021-07-13 194bf12
Numpy docstring format for warnings
