### 2021-07-13 cea1f02
DOC: Inverse transform sampling accuracy warning (#472)
