+++ skypy/utils/random.py
@@ -41,6 +41,13 @@ def schechter(alpha, x_min, x_max, resolution=100, size=None, scale=1.):
     x_sample : array_like
         Samples drawn from the Schechter function.
 
+    Warnings
+    --------
+    The inverse cumulative distribution function is approximated from the
+    Schechter function evaluated on a logarithmically-spaced grid. The user
+    must choose the `resolution` of this grid to satisfy their desired
+    numerical accuracy.
+
     References
     ----------
     .. [1] https://en.wikipedia.org/wiki/Luminosity_function_(astronomy)
