### 2021-07-19 c5e1007
initial commit of artale models
