+++ skypy/gravitational_waves/merger_rate.py
@@ -5,11 +5,13 @@ for individual galaxies.
 
 """
 
+import numpy as np
 from astropy import constants, units
 
 
 __all__ = [
     'b_band_merger_rate',
+    'm_star_merger_rate'
 ]
 
 
@@ -29,6 +31,214 @@ abadie_table_III = {
         'high': 20}
 }
 
+artale_tables = {
+    'NS-NS' : {
+            'redshift': [0.1, 1.0, 2.0, 6.0],
+            'alpha1': [1.038, 1.109, 1.050, 1.027],
+            'alpha1_err': [0.001, 0.001, 0.001, 0.003],
+            'alpha2': [-6.09, -6.214, -5.533, -5.029],
+            'alpha2_err': [0.010, 0.006, 0.006, 0.021],
+            'beta1': [0.800, 0.964, 0.977, 1.113],
+            'beta1_err': [0.002, 0.001, 0.001, 0.003],
+            'beta2': [0.323, 0.155, 0.068, -0.070],
+            'beta2_err': [0.002, 0.001, 0.001, 0.002],
+            'beta3': [-3.555, -4.819, -4.874, -5.764],
+            'beta3_err': [0.018, 0.013, 0.011, 0.026],
+            'gamma1': [0.701, 0.896, 1.018, 1.137],
+            'gamma1_err': [0.002, 0.002, 0.002, 0.004],
+            'gamma2': [0.356, 0.184, 0.048, -0.082],
+            'gamma2_err': [0.002, 0.001, 0.001, 0.002],
+            'gamma3': [0.411, 0.222, -0.103, -0.053],
+            'gamma3_err': [0.005, 0.003, 0.002, 0.004],
+            'gamma4': [-1.968, -3.795, -5.451, -6.104],
+            'gamma4_err': [0.026, 0.019, 0.017, 0.037],
+    }
+    'NS-BH' : {
+            'redshift': [0.1, 1.0, 2.0, 6.0],
+            'alpha1': [0.824, 0.873, 0.913, 0.965],
+            'alpha1_err': [0.001, 0.001, 0.001, 0.002],
+            'alpha2': [-4.731, -4.478, -4.401, -4.315],
+            'alpha2_err': [0.008, 0.008, 0.007, 0.018],
+            'beta1': [0.711, 0.813, 0.871, 0.985],
+            'beta1_err': [0.002, 0.002, 0.002, 0.003],
+            'beta2': [0.150, 0.064, 0.039, -0.017],
+            'beta2_err': [0.002, 0.002, 0.001, 0.001],
+            'beta3': [-3.536, -3.900, -4.019, -4.490],
+            'beta3_err': [0.016, 0.018, 0.014, 0.024],
+            'gamma1': [0.833, 1.074, 1.084, 0.978],
+            'gamma1_err': [0.002, 0.002, 0.002, 0.003],
+            'gamma2': [0.101, -0.058, -0.068, -0.013],
+            'gamma2_err': [0.002, 0.002, 0.001, 0.002],
+            'gamma3': [-0.461, -0.788, -0.535, 0.016],
+            'gamma3_err': [0.004, 0.004, 0.003, 0.004],
+            'gamma4': [-5.434, -7.733, -7.055, -4.386],
+            'gamma4_err': [0.022, 0.023, 0.018, 0.034],
+    }
+    'BH-BH' : {
+            'redshift': [0.1, 1.0, 2.0, 6.0],
+            'alpha1': [0.807, 0.813, 0.831, 0.933],
+            'alpha1_err': [0.001, 0.001, 0.001, 0.004],
+            'alpha2': [-4.310, -3.845, -3.600, -4.190],
[…]
