### 2021-07-19 e8437f9
fixed doc example, MSun units
