### 2021-07-19 c2ef42e
added sfr and metallicity model docs and tests
