+++ skypy/gravitational_waves/__init__.py
@@ -11,6 +11,7 @@ Merger Rates
 
    b_band_merger_rate
    m_star_merger_rate
+   m_star_sfr_merger_rate
 
 """
 
