+++ setup.cfg
@@ -48,7 +48,7 @@ testpaths = "skypy" "docs"
 astropy_header = true
 doctest_plus = enabled
 text_file_format = rst
-addopts = --doctest-rst
+#addopts = --doctest-rst
 
 [coverage:run]
 omit =
