### 2021-07-19 8d23946
put back doctest
