### 2021-07-21 cc259b9
added metallicity function to docs autosummary
