### 2021-07-22 019ddf3
Add intro text and shuffle import command lines
