+++ examples/galaxies/plot_ellipticity.py
@@ -9,8 +9,8 @@ in SkyPy.
 
 
 # %%
-# 3D ellipticity istribution
-# --------------------------
+# 3D Ellipticity Distribution
+# ---------------------------
 #
 # In Ryden 2004 [1]_, the ellipticity is sampled by randomly projecting
 # a 3D ellipsoid with principal axes :math:`A > B > C`.
@@ -23,17 +23,19 @@ in SkyPy.
 # with mean :math:`\mu` and standard deviation :math:`\sigma`.
 #
 #
-# Ellipticity SDSS Data
-# ---------------------
+# In ``SkyPy``, we can sample 3D ellipticities using the
+# :func:`skypy.galaxies.morphology.ryden04_ellipticity()` distribution
+# described above [1]_.
 #
-# Here we reproduce Figure 1 from [1]_, comparing our simulated galaxy
-# ellipticities against observational data from SDSS DR1.
+# Validation plot with SDSS Data
+# ------------------------------
+#
+# Here we validate our function by comparing our ``SkyPy`` simulated galaxy
+# ellipticities against observational data from SDSS DR1 in Figure 1 [1]_.
 # You can download the data file
-# :download:`SDSS_ellipticity <../../../examples/galaxies/SDSS_ellipticity.txt>`.
+# :download:`SDSS_ellipticity <../../../examples/galaxies/SDSS_ellipticity.txt>`,
 # stored as a 2D array: ellipticity_X, ellipticity_T.
 #
-# The best fit parameters [1]_:
-# :math:`\mu_\gamma, \sigma_\gamma, \mu, \sigma = 0.222, 0.057, -1.85, 0.89`.
 
 import numpy as np
 
@@ -43,17 +45,22 @@ Ngal = len(eX)
 e = np.hypot(eX, eT)
 q_amSDSS = np.sqrt((1 - e)/(1 + e))
 
-# Best fit parameters from Fig. 1 in Ryden 2004
-mu_gamma, sigma_gamma, mu, sigma = 0.222, 0.057, -1.85, 0.89
-
 
 # %%
-# SkyPy Ellipticity model
-# -----------------------
 #
-# We use the Ryden04 model in SkyPy.
+# In this example, we generate 100 galaxy ellipticity samples using the
+# ``SkyPy`` function
+# :func:`skypy.galaxies.morphology.ryden04_ellipticity()` and the
+# best fit parameters
+# given in Ryden 2004 [1]_:
+# :math:`\mu_\gamma =0.222`, :math:`\sigma_\gamma=0.057`, :math:`\mu =-1.85`
+# and :math:`\sigma=0.89`.
+#
 
 from skypy.galaxies.morphology import ryden04_ellipticity
+
+# Best fit parameters from Fig. 1 in Ryden 2004
+mu_gamma, sigma_gamma, mu, sigma = 0.222, 0.057, -1.85, 0.89
 
 # Binning scheme of Fig. 1
 bins = np.linspace(0, 1, 41)
@@ -62,6 +69,9 @@ mid = 0.5 * (bins[:-1] + bins[1:])
 # Mean and variance of sampling
 mean = np.zeros(len(bins)-1)
 var = np.zeros(len(bins)-1)
+
+# %%
+#
 
 # Create 100 SkyPy realisations
 for i in range(100):
@@ -82,9 +92,13 @@ for i in range(100):
 var = var/i
 std = np.sqrt(var)
 
+
 # %%
-# Plot
-# ----
+#
+# We now plot the distribution of axis ratio :math:`𝑞_{am}`
+# using adaptive moments in the i band, for exponential galaxies in the SDSS DR1
+# (solid line). The data points with error bars represent
+# the `SkyPy` simulation:
 #
 
 import matplotlib.pyplot as plt
