### 2021-07-22 94d16d3
Update skypy/gravitational_waves/tests/test_merger_rate.py
Co-authored-by: philipp128 <48715661+philipp128@users.noreply.github.com>
