### 2021-07-22 31e69f8
Update skypy/gravitational_waves/merger_rate.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
