+++ skypy/gravitational_waves/merger_rate.py
@@ -129,7 +129,7 @@ def m_star_merger_rate(redshift,
 
     Returns
     -------
-    merger_rate : array_like
+    merger_rate : (ngal,) array_like
         Merger rates for the galaxies in units of Gigayear^-1
 
     Notes
