### 2021-07-22 e8afd1d
Update skypy/gravitational_waves/merger_rate.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
