+++ skypy/gravitational_waves/merger_rate.py
@@ -115,7 +115,7 @@ def m_star_merger_rate(redshift,
 
     Parameters
     ----------
-    redshift : (ngal,) array-like
+    redshift : (ngal,) array_like
         The redshifts of the galaxies to generate merger
         rates for.
     m_star : (ngal,) array_like
