### 2021-07-22 c713d81
Update skypy/gravitational_waves/merger_rate.py
Co-authored-by: Lucia F. de la Bella <55983939+Lucia-Fonseca@users.noreply.github.com>
