### 2021-07-22 15f951d
removed private functions, doc examples
