### 2021-07-22 136ca7b
Add returns section to ryden04 function
