### 2021-07-22 587eaa3
Fix size of array
