### 2021-07-22 a18c050
Change function description
