### 2021-07-22 c47c5e8
Change load data commands to adapt to npz files
