+++ paper.bib
@@ -44,6 +44,42 @@
     adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Galsim2015,
+    author = {{Rowe}, B.~T.~P. and {Jarvis}, M. and {Mandelbaum}, R. and {Bernstein}, G.~M. and {Bosch}, J. and {Simet}, M. and {Meyers}, J.~E. and {Kacprzak}, T. and {Nakajima}, R. and {Zuntz}, J. and {Miyatake}, H. and {Dietrich}, J.~P. and {Armstrong}, R. and {Melchior}, P. and {Gill}, M.~S.~S.},
+    title = "{GALSIM: The modular galaxy image simulation toolkit}",
+    journal = {Astronomy and Computing},
+    keywords = {Methods: data analysis, Techniques: image processing, Gravitational lensing, Cosmology: observations, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, 85-04},
+    year = 2015,
+    month = apr,
+    volume = {10},
+    pages = {121-150},
+    doi = {10.1016/j.ascom.2015.02.002},
+    archivePrefix = {arXiv},
+    eprint = {1407.7676},
+    primaryClass = {astro-ph.IM},
+    adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&C....10..121R},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{halotools2017,
+    author = {{Hearin}, Andrew P. and {Campbell}, Duncan and {Tollerud}, Erik and {Behroozi}, Peter and {Diemer}, Benedikt and {Goldbaum}, Nathan J. and {Jennings}, Elise and {Leauthaud}, Alexie and {Mao}, Yao-Yuan and {More}, Surhud and {Parejko}, John and {Sinha}, Manodeep and {Sip{\"o}cz}, Brigitta and {Zentner}, Andrew},
+    title = "{Forward Modeling of Large-scale Structure: An Open-source Approach with Halotools}",
+    journal = {The Astronomical Journal},
+    keywords = {cosmology: theory, galaxies: halos, galaxies: statistics, large-scale structure of universe, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+    year = 2017,
+    month = nov,
+    volume = {154},
+    number = {5},
+    eid = {190},
+    pages = {190},
+    doi = {10.3847/1538-3881/aa859f},
+    archivePrefix = {arXiv},
+    eprint = {1606.04106},
+    primaryClass = {astro-ph.IM},
+    adsurl = {https://ui.adsabs.harvard.edu/abs/2017AJ....154..190H},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{LIGO2015,
     author = {{LIGO Scientific Collaboration}},
     title = "{Advanced LIGO}",
@@ -209,5 +245,21 @@
     eprint = {1807.06205},
     primaryClass = {astro-ph.CO},
     adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...641A...1P},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{popsynth2021,
+    author = {{Burgess}, J. and {Capel}, Francesca},
+    title = "{popsynth: A generic astrophysical population synthesis framework}",
+    journal = {The Journal of Open Source Software},
[…]
