+++ paper.md
@@ -30,6 +30,9 @@ authors:
   - name: Ian W. Harry
     orcid: 0000-0002-5304-9372
     affiliation: 1
+  - name: William G. Hartley
+    orcid: 0000-0002-5304-9372
+    affiliation: 6
   - name: Coleman Krawczyk
     orcid: 0000-0001-9233-2341
     affiliation: 1
@@ -38,28 +41,28 @@ authors:
     affiliation: 1
   - name: Brian Nord
     orcid: 0000-0001-6706-8972
-    affiliation: "6, 7, 8"
+    affiliation: "7, 8, 9"
   - name: Laura K. Nuttall
     orcid: 0000-0002-8599-8791
     affiliation: 1
   - name: Richard P. Rollins^[richard.rollins\@ed.ac.uk]
     orcid: 0000-0003-1291-1023
-    affiliation: "9, 2"
+    affiliation: "10, 2"
   - name: Philipp Sudek
     orcid: 0000-0001-8685-2308
     affiliation: 1
   - name: Sut-Ieng Tam
     orcid: 0000-0002-6724-833X
-    affiliation: 10
+    affiliation: 11
   - name: Nicolas Tessore
     orcid: 0000-0002-9696-7931
-    affiliation: 11
+    affiliation: 12
   - name: Arthur E. Tolley
     orcid: 0000-0001-9841-943X
     affiliation: 1
   - name: Keiichi Umetsu
     orcid: 0000-0002-7196-4822
-    affiliation: 10
+    affiliation: 11
   - name: Andrew R. Williamson
     orcid: 0000-0002-7627-8688
     affiliation: 1
@@ -77,20 +80,22 @@ affiliations:
     index: 4
   - name: Department of Physics, University of Oxford
     index: 5
-  - name: Fermi National Accelerator Laboratory
+  - name: Department of Astronomy, University of Geneva
     index: 6
-  - name: Kavli Institute for Cosmological Physics, University of Chicago
+  - name: Fermi National Accelerator Laboratory
     index: 7
-  - name: Department of Astronomy and Astrophysics, University of Chicago
+  - name: Kavli Institute for Cosmological Physics, University of Chicago
     index: 8
-  - name: Institute for Astronomy, University of Edinburgh
+  - name: Department of Astronomy and Astrophysics, University of Chicago
     index: 9
-  - name: Institute of Astronomy and Astrophysics, Academia Sinica
+  - name: Institute for Astronomy, University of Edinburgh
     index: 10
-  - name: Department of Physics and Astronomy, University College London
+  - name: Institute of Astronomy and Astrophysics, Academia Sinica
     index: 11
+  - name: Department of Physics and Astronomy, University College London
+    index: 12
 
-date: 30 June 2021
+date: 31 August 2021
 bibliography: paper.bib
 
 ---
@@ -103,7 +108,7 @@ Version `0.4` implements functions that model various properties of galaxies inc
 
 # Statement of need
 
[…]
