### 2021-08-31 skypyproject/skypy#478 @rrjbca
ENH: Rename noise keyword argument in redshifts_from_comoving_density function
It was noted in the review for our JOSS paper that the keyword argument `noise` for the function `redshifts_from_comoving_density` has a misleading name. See the review comment [here](https://github.com/openjournals/joss-reviews/issues/3056#issuecomment-877270540). They suggested renaming to `fixed_N`. My own suggestion would be `shot_noise` (or perhaps `poisson`). Since this would require a breaking API change, this should only be made as part of a future major release (if at all).
