### 2021-08-31 #3056 @rrjbca
Thank you for the reminder @arfon (and my apologies for missing your last message and not responding). @cescalara please find our response to your additional comments below:
- We have added a sentence to the Statement of Need specifically referring to the software packages that you mentioned in your review (GalSim, Halotools and popsynth).
- We agree that sampling redshifts and luminosities with an efficient rejection sampling method would be preferable. However, it would require reasonable proposal distributions which to our knowledge is an open problem. For Schechter luminosity distributions, left-truncated gamma distributions have been investigated, but only for positive indices and restricted scales. Furthermore, proposal distributions for redshifts would need to depend on cosmology. The huge range of possible distribution shapes make a generic proposal distribution challenging. We have merged [PR #472](https://github.com/skypyproject/skypy/pull/472) adding warnings to the docstrings for `schechter` and `redshifts_from_comoving_density` highlighting that the user must choose the `resolution` and `redshift` parameters to satisfy their own numerical accuracy requirements.
- While we agree that `noise` was perhaps a poor choice for the name of the keyword argument you mention, regrettably it is now a stable API used by other projects downstream. The keyword is documented and we have opened [Issue #478](https://github.com/skypyproject/skypy/issues/478) to consider this change in a future major release.
