### 2021-08-31 #3056 @rrjbca
Hi @arfon a quick question. In our [.zenodo.json](https://github.com/skypyproject/skypy/blob/main/.zenodo.json) we currently have:

```
"title": "SkyPy",
"description": "A package for modelling the Universe",
```

When you say the zenodo title needs to match the paper title does that mean we need to change it to this:

```
"title": "SkyPy: A package for modelling the Universe",
```

or is what we have sufficient? You can see the zenodo record for our last release [here](https://zenodo.org/record/4475347)
