### 2021-08-31 #3056 @arfon
> or is what we have sufficient? You can see the zenodo record for our last release here

I think what you currently have is sufficient.
