### 2021-08-31 skypyproject/skypy#443 @rrjbca
Resolved by #456 and #460. Closing
