### 2021-09-01 8fc0544
Update dependencies to exclude astropy 4.3.x
