### 2021-09-03 ff54808
Set astropy latest support version to 4.3
