+++ tox.ini
@@ -3,7 +3,7 @@ envlist =
     py{36,37,38,39}-test{,-all}{,-dev,-latest,-oldest}{,-cov}
     py{36,37,38,39}-test-numpy{116,117,118,119,120}
     py{36,37,38,39}-test-scipy{12,13,14,15,16}
-    py{36,37,38,39}-test-astropy{40,41,42}
+    py{36,37,38,39}-test-astropy{40,41,42,43}
     build_docs
     linkcheck
     codestyle
@@ -51,6 +51,7 @@ description =
     astropy40: with astropy 4.0.*
     astropy41: with astropy 4.1.*
     astropy42: with astropy 4.2.*
+    astropy43: with astropy 4.3.*
 
 # The following provides some specific pinnings for key packages
 deps =
@@ -70,12 +71,13 @@ deps =
     astropy40: astropy==4.0.*
     astropy41: astropy==4.1.*
     astropy42: astropy==4.2.*
+    astropy43: astropy==4.3.*
 
     dev: :NIGHTLY:numpy
     dev: :NIGHTLY:scipy
     dev: git+https://github.com/astropy/astropy.git#egg=astropy
 
-    latest: astropy==4.2.*
+    latest: astropy==4.3.*
     latest: numpy==1.20.*
     latest: scipy==1.6.*
 
