+++ setup.cfg
@@ -17,7 +17,7 @@ packages = find:
 python_requires = >=3.6
 setup_requires = setuptools_scm
 install_requires =
-    astropy>=4,<4.3
+    astropy>=4
     networkx
     numpy
     scipy
