### 2021-09-03 acffc20
Add compatibility workflow badge
