### 2021-09-03 609b968
Set numpy latest support version to 1.21
