+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =
     py{36,37,38,39}-test{,-all}{,-dev,-latest,-oldest}{,-cov}
-    py{36,37,38,39}-test-numpy{116,117,118,119,120}
+    py{36,37,38,39}-test-numpy{116,117,118,119,120,121}
     py{36,37,38,39}-test-scipy{12,13,14,15,16}
     py{36,37,38,39}-test-astropy{40,41,42}
     build_docs
@@ -43,6 +43,7 @@ description =
     numpy118: with numpy 1.18.*
     numpy119: with numpy 1.19.*
     numpy120: with numpy 1.20.*
+    numpy121: with numpy 1.21.*
     scipy12: with scipy 1.2.*
     scipy13: with scipy 1.3.*
     scipy14: with scipy 1.4.*
@@ -60,6 +61,7 @@ deps =
     numpy118: numpy==1.18.*
     numpy119: numpy==1.19.*
     numpy120: numpy==1.20.*
+    numpy121: numpy==1.21.*
 
     scipy12: scipy==1.2.*
     scipy13: scipy==1.3.*
@@ -76,7 +78,7 @@ deps =
     dev: git+https://github.com/astropy/astropy.git#egg=astropy
 
     latest: astropy==4.2.*
-    latest: numpy==1.20.*
+    latest: numpy==1.21.*
     latest: scipy==1.6.*
 
     oldest: astropy==4.0.*
