### 2021-09-03 33e1b12
Set scipy latest support version to 1.7
