+++ tox.ini
@@ -2,7 +2,7 @@
 envlist =
     py{36,37,38,39}-test{,-all}{,-dev,-latest,-oldest}{,-cov}
     py{36,37,38,39}-test-numpy{116,117,118,119,120,121}
-    py{36,37,38,39}-test-scipy{12,13,14,15,16}
+    py{36,37,38,39}-test-scipy{12,13,14,15,16,17}
     py{36,37,38,39}-test-astropy{40,41,42}
     build_docs
     linkcheck
@@ -49,6 +49,7 @@ description =
     scipy14: with scipy 1.4.*
     scipy15: with scipy 1.5.*
     scipy16: with scipy 1.6.*
+    scipy17: with scipy 1.7.*
     astropy40: with astropy 4.0.*
     astropy41: with astropy 4.1.*
     astropy42: with astropy 4.2.*
@@ -68,6 +69,7 @@ deps =
     scipy14: scipy==1.4.*
     scipy15: scipy==1.5.*
     scipy16: scipy==1.6.*
+    scipy17: scipy==1.7.*
 
     astropy40: astropy==4.0.*
     astropy41: astropy==4.1.*
@@ -79,7 +81,7 @@ deps =
 
     latest: astropy==4.2.*
     latest: numpy==1.21.*
-    latest: scipy==1.6.*
+    latest: scipy==1.7.*
 
     oldest: astropy==4.0.*
     oldest: numpy==1.16.*
