### 2021-09-03 70a5ecd
Use tmp_path fixture for temporary files in unit tests
