+++ skypy/pipeline/tests/test_skypy.py
@@ -10,7 +10,7 @@ from skypy.pipeline._items import Call
 from skypy.pipeline.scripts import skypy
 
 
-def test_skypy():
+def test_skypy(tmp_path):
 
     # No arguments
     with pytest.raises(SystemExit) as e:
@@ -36,36 +36,39 @@ def test_skypy():
     assert e.value.code == 2
 
     # Process empty config file
-    filename = get_pkg_data_filename('data/empty_config.yml')
-    assert skypy.main([filename, 'empty.fits']) == 0
+    config_filename = get_pkg_data_filename('data/empty_config.yml')
+    output_filename = str(tmp_path / 'empty.fits')
+    assert skypy.main([config_filename, output_filename]) == 0
 
     # Process test config file
-    filename = get_pkg_data_filename('data/test_config.yml')
-    assert skypy.main([filename, 'test.fits']) == 0
+    config_filename = get_pkg_data_filename('data/test_config.yml')
+    output_filename = str(tmp_path / 'test.fits')
+    assert skypy.main([config_filename, output_filename]) == 0
 
     # Invalid file format
+    output_filename = str(tmp_path / 'test.invalid')
     with pytest.raises(SystemExit) as e:
-        skypy.main([filename, 'test.invalid'])
+        skypy.main([config_filename, output_filename])
     assert e.value.code == 2
 
 
-def test_logging(capsys):
+def test_logging(capsys, tmp_path):
 
     # Run skypy with default verbosity and check log is empty
-    filename = get_pkg_data_filename('data/test_config.yml')
-    output_file = 'logging.fits'
-    skypy.main([filename, output_file])
+    config_filename = get_pkg_data_filename('data/test_config.yml')
+    output_filename = str(tmp_path / 'logging.fits')
+    skypy.main([config_filename, output_filename])
     out, err = capsys.readouterr()
     assert(not err)
 
     # Run again with increased verbosity and capture log. Force an exception by
     # not using the "--overwrite" flag when the output file already exists.
     with pytest.raises(SystemExit):
-        skypy.main([filename, output_file, '--verbose'])
+        skypy.main([config_filename, output_filename, '--verbose'])
     out, err = capsys.readouterr()
 
     # Determine all DAG jobs and function calls from config
-    config = load_skypy_yaml(filename)
+    config = load_skypy_yaml(config_filename)
     cosmology = config.pop('cosmology', None)
     tables = config.pop('tables', {})
     config.update({k: v.pop('.init', Call(Table)) for k, v in tables.items()})
@@ -88,21 +91,13 @@ def test_logging(capsys):
         assert("[INFO] skypy.pipeline: Setting cosmology" in err)
 
     # Check writing output file is in the log
-    assert(f"[INFO] skypy: Writing {output_file}" in err)
+    assert(f"[INFO] skypy: Writing {output_filename}" in err)
 
     # Check error for existing output file is in the log
-    assert(f"[ERROR] skypy: File '{output_file}' already exists." in err)
+    assert(f"[ERROR] skypy: File '{output_filename}' already exists." in err)
 
     # Run again with decreased verbosity and check the log is empty
     with pytest.raises(SystemExit):
-        skypy.main([filename, output_file, '-qq'])
+        skypy.main([config_filename, output_filename, '-qq'])
     out, err = capsys.readouterr()
     assert(not err)
-
-
-def teardown_module(module):
-
-    # Remove fits file generated in test_skypy
-    os.remove('empty.fits')
-    os.remove('logging.fits')
-    os.remove('test.fits')
