### 2021-09-03 41fe6da
flake8: remove unused os imports
