### 2021-09-06 296d750
Modify text for clarity and delete link to reference
