+++ skypy/galaxies/morphology.py
@@ -254,7 +254,7 @@ def ryden04_ellipticity(mu_gamma, sigma_gamma, mu, sigma, size=None):
     Returns
     -------
     ellipticity: (size,) array_like
-        Ellipticity distribution sampled from the Ryden 2004 model [1].
+        Ellipticities sampled from the Ryden 2004 model.
 
     References
     ----------
