### 2021-09-06 3cfb5d5
Modify unit test to match OSError raised by astropy
