### 2021-09-06 76c36e8
MAINT: Update numpy and scipy latest supported versions (#488)
