### 2021-09-06 e9dc851
Use clearer variables names
