+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
 
-requires = ["setuptools",
+requires = ["setuptools<=58.0.1",
             "setuptools_scm",
             "wheel"]
 
