+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
 
-requires = ["setuptools<=58.0.1",
+requires = ["setuptools",
             "setuptools_scm",
             "wheel"]
 
