### 2021-09-07 40456ee
DEV: setuptools==58.0.0 (#493)
Co-authored-by: Nicolas Tessore <n.tessore@ucl.ac.uk>
