+++ docs/requirements.txt
@@ -1,0 +1,1 @@
+setuptools==58.0.0
