### 2021-09-07 1419e08
Revert "DEV: setuptools==58.0.0 (#493)"
This reverts commit 40456ee178fad5087f7cbbc103ca2856d948b166.
