+++ .readthedocs.yml
@@ -9,7 +9,6 @@ sphinx:
 python:
   version: 3.7
   install:
-    - requirements: docs/requirements.txt
     - method: pip
       path: .
       extra_requirements:
