### 2021-09-08 aa95dec
DEV: Enable pip to install pre-releases in the tox dev environments (#491)
