+++ tox.ini
@@ -85,6 +85,10 @@ extras =
     test
     all: all
 
+# Enable pip to install pre-releases in the dev environment
+pip_pre =
+    dev: True
+
 commands =
     pip freeze
     !cov: pytest --pyargs skypy {toxinidir}/docs {posargs}
