### 2021-09-08 c2cb389
TST: Use tmp_path fixture for temporary files in unit tests (#489)
