### 2021-09-08 83cb5cd
Update skypy/pipeline/tests/test_pipeline.py
