### 2021-09-10 3850e59
BLD: Set astropy latest supported version to 4.3 and speclite minversion to 0.14 (#486)
